fix: critical double-free crash in api_file_list_handler
Root cause: path and path_param pointed to same memory when
get_query_param returned non-NULL. free(path) then free(path_param)
caused 'block already marked as free' assert in tlsf_free.

Also fix %zu format specifiers in file_manager.c (not supported on ESP32).

+++ main/web_server/file_manager.c
@@ -144,7 +144,7 @@ esp_err_t file_manager_upload(const char *path, const uint8_t *data, size_t len)
         return ESP_ERR_INVALID_ARG;
     }
 
-    ESP_LOGI(TAG, "Uploading file: %s (%zu bytes)", path, len);
+    ESP_LOGI(TAG, "Uploading file: %s (%lu bytes)", path, (unsigned long)len);
 
     // Ensure parent directory exists
     char parent_dir[512];
@@ -166,7 +166,7 @@ esp_err_t file_manager_upload(const char *path, const uint8_t *data, size_t len)
     fclose(file);
 
     if (written != len) {
-        ESP_LOGE(TAG, "Write failed: wrote %zu of %zu bytes", written, len);
+        ESP_LOGE(TAG, "Write failed: wrote %lu of %lu bytes", (unsigned long)written, (unsigned long)len);
         return ESP_ERR_INVALID_SIZE;
     }
